requirements.txt
deepeval>=2.0.0
streamlit>=1.30.0
openai>=1.10.0
python-dotenv>=1.0.0
pandas>=2.0.0
